Moving mixnode bonding queries into their own file
