+++ contracts/mixnet/src/gateways/mod.rs
@@ -1,1 +1,2 @@
+pub mod queries;
 pub mod transactions;
