+++ contracts/mixnet/src/mixnodes/mod.rs
@@ -1,2 +1,4 @@
+pub mod bonding_queries;
 pub mod bonding_transactions;
+pub mod delegation_queries;
 pub mod delegation_transactions;
