+++ contracts/mixnet/src/rewards/queries.rs
@@ -1,0 +1,7 @@
+pub(crate) fn query_reward_pool(deps: Deps) -> Uint128 {
+    reward_pool_value(deps.storage)
+}
+
+pub(crate) fn query_circulating_supply(deps: Deps) -> Uint128 {
+    circulating_supply(deps.storage)
+}
